Upgrade pysignalr to version 1.3.0
pysmarlaapi is now included in Homeassistant Docker images by default.
Adding it caused the websockets library to be downgraded from version
15 to version 13, which is problematic for other things.

This is caused by pysmarlaapi requiring pysignalr version 1.1.0 which then
has a hard dependency on websockets version 13. The latest version of
pysignalr is using websockets 15, so upgrading pysmarlaapi to use that
instead fixes the problem.

Since this project does not have any unit tests and I don't have the
hardware to test this with (I don't actually even know what that hardware
might be) this PR is entirely untested.